add dialect trigger options
