+++ Tests/LinuxMain.swift
@@ -1,0 +1,2 @@
+#error("Please test with `swift test --enable-test-discovery`")
+
