+++ docker-compose.yml
@@ -1,22 +1,25 @@
-version: '3'
+version: '3.7'
+
+x-common: &common
+    environment:
+        MYSQL_ALLOW_EMPTY_PASSWORD: "true"
+        MYSQL_DATABASE: vapor_database
+        MYSQL_USER: vapor_username
+        MYSQL_PASSWORD: vapor_password
+    ports:
+      - 3306:3306
 
 services:
-  mysql-8:
-    image: mysql:8.0
-    environment:
-      MYSQL_ALLOW_EMPTY_PASSWORD: "true"
-      MYSQL_DATABASE: vapor_database
-      MYSQL_USER: vapor_username
-      MYSQL_PASSWORD: vapor_password
-    ports:
-      - 3306:3306
-  mysql-5:
-    image: mysql:5.7
-    environment:
-      MYSQL_ALLOW_EMPTY_PASSWORD: "true"
-      MYSQL_DATABASE: vapor_database
-      MYSQL_USER: vapor_username
-      MYSQL_PASSWORD: vapor_password
-    ports:
-      - 3306:3306
+    mysql:
+        image: mysql
+        <<: *common
+    maria:
+        image: mariadb
+        <<: *common
+    mysql-8:
+        image: mysql:8.0
+        <<: *common
+    mysql-5:
+        image: mysql:5.7
+        <<: *common
